rearrange and rename files
