+++ generate_saliency_maps.py
@@ -2,13 +2,10 @@
 Generate saliency maps for a model.
 
 Author: Andrew Justin (andrewjustinwx@gmail.com)
-Script version: 2024.8.3
+Script version: 2024.10.10
 """
-
 import argparse
-import sys
 import os
-sys.path.append(os.path.abspath(os.path.join(os.path.dirname(__file__), os.pardir)))  # this line allows us to import scripts outside the current directory
 from utils import data_utils
 import numpy as np
 import pandas as pd
