--- webapp.py
@@ -1,5 +1,0 @@
-import streamlit as st
-
-st.set_page_config(layout='wide')
-
-st.html("""<h1 style='color:aqua; text-align:center;'>AI2ES Fronts Project</h1>""")
